requirements.txt
blinker==1.6.2
click==8.1.3
colorama==0.4.6
et-xmlfile==1.1.0
Flask==2.3.2
importlib-metadata==6.7.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
MarkupSafe==2.1.3
numpy==1.24.3
openpyxl==3.1.2
pandas==2.0.2
python-dateutil==2.8.2
pytz==2023.3
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
threadpoolctl==3.1.0
tzdata==2023.3
Werkzeug==2.3.6